Support for decimal numbers

+++ src/main/java/com/kevinhe99/Calculator.java
@@ -2,19 +2,19 @@ package com.kevinhe99;
 
 public class Calculator {
 
-    public static Result add(int val1, int val2) {
+    public static Result add(double val1, double val2) {
         return new Result(val1 + val2, true);
     }
 
-    public static Result subtract(int val1, int val2) {
+    public static Result subtract(double val1, double val2) {
         return new Result(val1 - val2, true);
     }
 
-    public static Result multiply(int val1, int val2) {
+    public static Result multiply(double val1, double val2) {
         return new Result(val1 * val2, true);
     }
 
-    public static Result divide(int val1, int val2) {
+    public static Result divide(double val1, double val2) {
         if (val2 == 0) {
             return new Result(0, false);
         }
